Basic implementation of the special query functions
I replaced the vector that holds the messages with a map with a key of
an `unsigned long long int` which holds the time the message was sent,
so the map will automatically sort the messages based on the time and I
don't have to worry about any algorithms.

Following that, I have also implemented basic implementation for the
query after, before, and since functions, and since I made the change to
a map I don't have to do anything special with them, just insert and
they will be sorted.

In addition to that I also changed the access for the `nlohmann::json`
objects from standard array access operators (e.g. []) to the `.at()`
function which has boundary detection.

+++ include/Message.h
@@ -135,10 +135,10 @@ namespace GroupMe {
             /**
              * @brief Gets when the message was created at
              *
-             * @return unsigned int
+             * @return unsigned long long int
              *
              */
-            unsigned int getCreatedAt() const;
+            unsigned long long int getCreatedAt() const;
 
             /**
              * @brief Sets when the message was created at
@@ -146,7 +146,7 @@ namespace GroupMe {
              * @param createdAt The new time to set
              *
              */
-            void setCreatedAt(unsigned int createdAt);
+            void setCreatedAt(unsigned long long int createdAt);
 
             /**
              * @brief Gets the text attached to the message
@@ -255,11 +255,11 @@ namespace GroupMe {
 
             std::string m_guid;
 
-            unsigned int m_createdAt;
+            unsigned long long int m_createdAt;
 
             bool m_pinned;
 
-            unsigned int m_pinnedAt;
+            unsigned long long int m_pinnedAt;
 
             std::shared_ptr<GroupMe::User> m_pinnedBy;
